Optionally return a user-supplied default value
If the user chooses an invalid string (something not corresponding to the
`#to_s` of one of the objects), `Fuzz::Selector#pick` returns `nil`.

To make it easier to implement the null object pattern, it'd be nice if we could
return a better, user-supplied default value. This adds a `default:` option to
the `Fuzz::Selector` constructor.